+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Typescript Project Template
 
+## [v1.2.2]
+
+### Fixes
+- [TSTEMPLATE-7] Source mapping references wrong line numbers when run in dev/ts-node
+- [TSTEMPLATE-8] Set stack trace limit to Infinity
+- [TSTEMPLATE-9] Filter stack traces for "just my code"
+- [TSTEMPLATE-10] Remove webpack from default build-chain
+- [TSTEMPLATE-11] Remove dist references from package.json
+- [TSTEMPLATE-12] Add lint rule for unresolved promises
+
 ## [v1.2.1]
 
 ### Fixes